+++ packages/storage/src/providers/s3/apis/internal/list.ts
@@ -243,6 +243,7 @@ const _listWithPath = async ({
 	if (!contents) {
 		return {
 			items: [],
+			nextToken: nextContinuationToken,
 			excludedSubpaths,
 		};
 	}